refactor: move user profile under admin folder
